requirements.txt
fastapi
uvicorn
pydantic
numpy
onnx
onnxruntime-gpu
scikit-image
moviepy==1.0.3
python-multipart
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy<2.0
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.15.2
six==1.17.0
sympy==1.13.3
typing_extensions==4.13.2
sqlalchemy>=2.0.0
psycopg2-binary
databases
asyncpg
boto3==1.26.59
pytz
pynvml